\name{austin-package}
\alias{austin-package}
\alias{austin}
\docType{package}
\title{Austin}
\description{
  Austin helps you see what people do with words.
  Currently that means positioning documents on a presumed underlying
  policy scale using only word frequencies and
  some heroic assumptions about language generation.
}
\details{
  Currently implemented models are: Wordfish (\code{wordfish}),
  and new Wordscores (\code{wordscores}).
}
\author{
Will Lowe

Maintainer: Will Lowe <conjugateprior@gmail.com>
}
\references{}
%\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{}
