\docType{package}
\name{austin}
\alias{austin}
\alias{austin-package}
\alias{package-austin}
\title{Austin does things with words}
\description{
  Austin implements Wordfish, Wordscores, and related text
  scaling models. It also provides utility function for
  working with various sparse data formats used in text
  analysis, e.g. LDA-C, Matrix Market, and variants on the
  SVMLite format.
}
\author{
  Will Lowe \email{will.lowe@uni-mannheim.de}
}

