\name{coef.wordscores}
\alias{coef.wordscores}
\title{Wordscores from a wordscores model}
\usage{
  \method{coef}{wordscores} (object, ...)
}
\arguments{
  \item{object}{a wordscores model}

  \item{...}{Not used}
}
\value{
  A set of scores for words
}
\description{
  Wordscores from a wordscores model
}
\details{
  Just lists the wordscores
}
\author{
  Will Lowe
}

