\docType{data}
\name{demanif.foreign}
\alias{demanif.foreign}
\title{German party manifestos, foreign policy}
\description{
  A random sample of words from sections of German party
  manifestos concerning foreign policy issues from
  1990-2005.
}
\author{
  Will Lowe, from data collected by Slapin and Proksch
}
\references{
  J. Slapin and S.-O. Proksch (2008) 'A scaling model for
  estimating time-series party positions from texts'
  American Journal of Political Science 52(3), 705-722
}
\keyword{data}
\keyword{datasets}

