\docType{data}
\name{demanif}
\alias{demanif}
\title{German party manifestos}
\description{
  A random sample of words from German party manifestos
  from 1990-2005.
}
\author{
  Will Lowe, from data collected by Slapin and Proksch
}
\references{
  J. Slapin and S.-O. Proksch (2008) 'A scaling model for
  estimating time-series party positions from texts'
  American Journal of Political Science 52(3), 705-722
}
\keyword{data}
\keyword{dataset}

