\name{fitted.wordfish}
\alias{fitted.wordfish}
\title{Fitted values for a wordfish model}
\usage{
  \method{fitted}{wordfish} (object, ...)
}
\arguments{
  \item{object}{fitted wordfish model}

  \item{...}{Not used}
}
\value{
  matrix of expected counts
}
\description{
  Fitted values for a wordfish model
}
\details{
  The function generates a matrix of expected counts the
  same size as the data matrix used to fit the model.
}
\author{
  Will Lowe
}

