\docType{data}
\name{interestgroups}
\alias{interestgroups}
\title{Interest groups and the European Commission}
\description{
  Word counts from interest group statements and two
  versions of the European commission's proposals to reduce
  CO2 emmissions, from 2007.
}
\details{
  \code{comm1} and \code{comm2} are the commission's
  proposal before and after the proposals of the interest
  groups respectively.
}
\author{
  Heike Kluever
}
\references{
  H. Kluever (2009) 'Measuring influence group influence
  using quantitative text analysis' European Union Politics
  11:1.
}
\keyword{data}
\keyword{dataset}

