\docType{data}
\name{lbg}
\alias{lbg}
\title{Fake wordscores data}
\description{
  Fake data for showing the wordscores algorithm.
  Documents R1-R5 are 'reference' with scores: -1.5, -0.75,
  0, 0.75, 1.5 and V1 is 'virgin' with true score 0.45.
}
\author{
  Michael Laver, Ken Benoit, and John Garry
}
\references{
  Laver, M. Benoit, K. and Garry, J (2003) Extracting
  policy positions from political texts using words as
  data. American Political Science Review 97:2
}
\keyword{data}
\keyword{datasets}

