\name{print.summary.wordfish}
\alias{print.summary.wordfish}
\title{Print a wordfish model summary}
\usage{
  \method{print}{summary.wordfish} (x,
    digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
  \item{x}{wordfish model summary}

  \item{digits}{how many significant digits to use}

  \item{...}{Not used}
}
\value{
  None. Used for printing side-effect
}
\description{
  Print a wordfish model summary
}
\details{
  Prints a wordfish model summary
}
\author{
  Will Lowe
}

