\name{summary.wordscores}
\alias{summary.wordscores}
\title{Summarise a wordscores model}
\usage{
  \method{summary}{wordscores} (object, ...)
}
\arguments{
  \item{object}{wordscores model}

  \item{...}{Not used}
}
\value{
  A data.frame with document statistics
}
\description{
  Summarise a wordscores model
}
\details{
  Generates some minimal statistics about the document set
  from which wordscores have been calculated.
}
\author{
  Will Lowe
}

