\docType{data}
\name{ukmanif}
\alias{ukmanif}
\title{UK party manifestos}
\description{
  Word counts from UK party manifestos, in 1992 and 1997.
}
\author{
  Michael Laver, Ken Benoit, and John Garry
}
\references{
  Laver, M. Benoit, K. and Garry, J (2003) Extracting
  policy positions from political texts using words as
  data. American Political Science Review 97:2
}
\keyword{data}
\keyword{datasets}

